[3802] Setup state & client-test modules and update UT imports
