+++ settings.gradle
@@ -6,8 +6,10 @@ include (
 		':stream-chat-android-ui-common',
 		':stream-chat-android-ui-components',
 		':stream-chat-android-ui-components-sample',
+		':stream-chat-android-state',
 		':stream-chat-android-offline',
 		':stream-chat-android-client',
+		':stream-chat-android-client-test',
 		':stream-chat-android-core',
 		':stream-chat-android-test',
 		':stream-chat-android-docs',
