+++ stream-chat-android-client-test/src/main/java/io/getstream/chat/android/client/test/SynchronizedCoroutineTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.offline
+package io.getstream.chat.android.client.test
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -27,13 +27,13 @@ import kotlinx.coroutines.test.runTest
  * possible.
  */
 @ExperimentalCoroutinesApi
-internal interface SynchronizedCoroutineTest {
+public interface SynchronizedCoroutineTest {
 
     /** Returns test scope. */
-    fun getTestScope(): TestScope
+    public fun getTestScope(): TestScope
 
     /** Helper function that synchronize test scope and run your test in another scope blocking. */
-    fun coroutineTest(block: suspend CoroutineScope.() -> Unit): Unit = runTest {
+    public fun coroutineTest(block: suspend CoroutineScope.() -> Unit): Unit = runTest {
         getTestScope().launch(block = block).join()
     }
 }
