+++ stream-chat-android-offline/build.gradle
@@ -57,6 +57,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 
 dependencies {
     api project(":stream-chat-android-client")
+    api project(":stream-chat-android-state")
 
     // Kotlin
     implementation Dependencies.kotlinReflect
@@ -77,6 +78,7 @@ dependencies {
 
     // Tests
     testImplementation project(":stream-chat-android-test")
+    testImplementation project(":stream-chat-android-client-test")
     testImplementation Dependencies.moshiKotlin
 
     testImplementation Dependencies.junitJupiterApi
