+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/events/TypingEventsTest.kt
@@ -19,10 +19,10 @@ package io.getstream.chat.android.offline.events
 import io.getstream.chat.android.client.models.EventType
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.plugin.listeners.TypingEventListener
+import io.getstream.chat.android.client.test.randomUser
 import io.getstream.chat.android.offline.plugin.listener.internal.TypingEventListenerImpl
 import io.getstream.chat.android.offline.plugin.state.StateRegistry
 import io.getstream.chat.android.offline.plugin.state.channel.internal.toMutableState
-import io.getstream.chat.android.offline.randomUser
 import io.getstream.chat.android.test.TestCoroutineExtension
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
