+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/internal/extensions/ChannelExtensionsTest.kt
@@ -27,9 +27,9 @@ import io.getstream.chat.android.client.extensions.internal.applyPagination
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.parser2.adapters.DateAdapter
+import io.getstream.chat.android.client.test.randomChannel
 import io.getstream.chat.android.offline.model.querychannels.pagination.internal.QueryChannelsPaginationRequest
 import io.getstream.chat.android.offline.model.querychannels.pagination.internal.toAnyChannelPaginationRequest
-import io.getstream.chat.android.offline.randomChannel
 import okio.buffer
 import okio.source
 import org.amshove.kluent.shouldBeEqualTo
