+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/querychannels/QueryChannelsSortTest.kt
@@ -19,7 +19,7 @@ package io.getstream.chat.android.offline.querychannels
 import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.CustomObject
-import io.getstream.chat.android.offline.randomChannel
+import io.getstream.chat.android.client.test.randomChannel
 import org.amshove.kluent.`should be equal to`
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
