+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/repository/ChannelConfigRepositoryTest.kt
@@ -19,8 +19,8 @@ package io.getstream.chat.android.offline.repository
 import com.flextrade.jfixture.JFixture
 import com.flextrade.kfixture.KFixture
 import io.getstream.chat.android.client.persistance.repository.ChannelConfigRepository
-import io.getstream.chat.android.offline.randomChannelConfig
-import io.getstream.chat.android.offline.randomConfig
+import io.getstream.chat.android.client.test.randomChannelConfig
+import io.getstream.chat.android.client.test.randomConfig
 import io.getstream.chat.android.offline.repository.domain.channelconfig.internal.ChannelConfigDao
 import io.getstream.chat.android.offline.repository.domain.channelconfig.internal.ChannelConfigEntity
 import io.getstream.chat.android.offline.repository.domain.channelconfig.internal.ChannelConfigInnerEntity
