+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/repository/ReactionRepositoryTest.kt
@@ -18,9 +18,9 @@ package io.getstream.chat.android.offline.repository
 
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.persistance.repository.ReactionRepository
+import io.getstream.chat.android.client.test.randomReaction
+import io.getstream.chat.android.client.test.randomUser
 import io.getstream.chat.android.client.utils.SyncStatus
-import io.getstream.chat.android.offline.randomReaction
-import io.getstream.chat.android.offline.randomUser
 import io.getstream.chat.android.offline.repository.domain.reaction.internal.DatabaseReactionRepository
 import io.getstream.chat.android.offline.repository.domain.reaction.internal.ReactionDao
 import io.getstream.chat.android.offline.repository.domain.reaction.internal.toEntity
