+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/repository/UserRepositoryTests.kt
@@ -19,7 +19,7 @@ package io.getstream.chat.android.offline.repository
 import app.cash.turbine.test
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.persistance.repository.UserRepository
-import io.getstream.chat.android.offline.randomUser
+import io.getstream.chat.android.client.test.randomUser
 import io.getstream.chat.android.offline.repository.domain.user.internal.DatabaseUserRepository
 import io.getstream.chat.android.offline.repository.domain.user.internal.UserDao
 import io.getstream.chat.android.test.TestCoroutineRule
