+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/repository/database/converter/MapConverterTest.kt
@@ -16,10 +16,10 @@
 
 package io.getstream.chat.android.offline.repository.database.converter
 
+import io.getstream.chat.android.client.test.utils.TestDataHelper
 import io.getstream.chat.android.offline.repository.database.converter.internal.MapConverter
 import io.getstream.chat.android.offline.repository.domain.channel.member.internal.MemberEntity
 import io.getstream.chat.android.offline.repository.domain.channel.userread.internal.ChannelUserReadEntity
-import io.getstream.chat.android.offline.utils.TestDataHelper
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.Test
 import java.util.Date
