+++ stream-chat-android-state/build.gradle
@@ -1,0 +1,104 @@
+import io.getstream.chat.android.Dependencies
+
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+}
+
+
+ext {
+    PUBLISH_GROUP_ID = io.getstream.chat.android.Configuration.artifactGroup
+    PUBLISH_ARTIFACT_ID = 'stream-chat-android-state'
+    PUBLISH_VERSION = rootVersionName
+}
+
+apply from: "${rootDir}/scripts/publish-module.gradle"
+apply from: "${rootDir}/scripts/android.gradle"
+
+android {
+    defaultConfig {
+        vectorDrawables.useSupportLibrary = true
+    }
+
+    resourcePrefix 'stream_'
+
+    testOptions {
+        unitTests {
+            includeAndroidResources = true
+            unitTests.returnDefaultValues = true
+            // Show the result of every unit test, even if it passes.
+            all {
+                testLogging {
+                    events 'passed', 'skipped', 'failed', 'standardOut', 'standardError'
+                }
+            }
+        }
+    }
+
+    sourceSets {
+        all {
+            it.java.srcDir "src/$it.name/kotlin"
+        }
+    }
+}
+
+
+tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+    kotlinOptions {
+        freeCompilerArgs += [
+                '-progressive',
+                '-Xexplicit-api=strict',
+                '-Xopt-in=io.getstream.chat.android.core.internal.InternalStreamChatApi'
+        ]
+    }
+}
+
+dependencies {
+    api project(":stream-chat-android-client")
+
+    // Kotlin
+    implementation Dependencies.kotlinReflect
+    implementation Dependencies.coroutinesCore
+
+    // Google libs
+    implementation Dependencies.androidxActivityKtx
+    implementation Dependencies.androidxAnnotations
+    implementation Dependencies.androidxAppCompat
+    implementation Dependencies.workRuntimeKtx
+
+    // Tests
+    testImplementation project(":stream-chat-android-test")
+    testImplementation project(":stream-chat-android-client-test")
+    testImplementation Dependencies.moshiKotlin
+
+    testImplementation Dependencies.junitJupiterApi
+    testImplementation Dependencies.junitJupiterParams
+    testRuntimeOnly Dependencies.junitJupiterEngine
+    testRuntimeOnly Dependencies.junitVintageEngine
+
+    testImplementation Dependencies.androidxTest
+    testImplementation Dependencies.androidxTestJunit
+    testImplementation Dependencies.androidxArchCoreTest
+    testImplementation Dependencies.androidxRecyclerview // for performance test
+    testImplementation Dependencies.roomTesting
+    testImplementation Dependencies.workTesting
+
+    testImplementation Dependencies.coroutinesTest
+
+    testImplementation Dependencies.robolectric
+    testImplementation Dependencies.dotenv
+
+    testImplementation Dependencies.kfixture
+    testImplementation Dependencies.kluent
+    testImplementation Dependencies.mockito
+    testImplementation Dependencies.mockitoKotlin
+    testImplementation Dependencies.turbine
+
+    // Instrumentation tests
+    androidTestImplementation Dependencies.androidxTestJunit
+    androidTestImplementation Dependencies.espressoCore
+    androidTestImplementation Dependencies.junit4
+    androidTestImplementation Dependencies.kluent
+    androidTestImplementation Dependencies.mockito
+    androidTestImplementation Dependencies.mockitoKotlin
+}
