+++ stream-chat-android-state/src/test/resources/robolectric.properties
@@ -1,0 +1,1 @@
+sdk=28
